refactor(#7): extract upcoming holidays to custom hook
